Added files via upload
Some first drafts of UML models, should do for the submission tomorrow. They differ slightly from the requirements document in places because I was trying to think about how exactly to implement all of this in Java. Files have been created using Violet UML Editor, it's free to download and is solidly okay. I was working on a sequence diagram but never got round to finishing it because of coursework this week. It shouldn't matter too much for the submission tomorrow but i'll work on it over the weekend and we should have it by next week.